Adds a pending subscription helper method
This pull request makes it easier to subscribe someone to a newsletter as pending, this will require the user to click a confirmation email send by mailchimp.

I have not included a helper for subscribe or update, this is stop the possibility of setting an already confirmed subscription back to pending. this could be added in the future if required.