Adds a pending subscription helper method
What it Does:
This pull request adds a new helper method to add a newsletter subscription while setting the status to pending, this will require the individual to confirm their subscription via email.

Side Note:
I have not included a helper for the subscribeOrUpdate method, this would introduce the possibility of setting an already confirmed subscription back to pending, if this is requested this could be accomplished safely by checking if the email is subscribed to the list before merging the options.