Update 'Using mailchimp' readme section
While using your package, I followed the instructions, but I was always getting empty responses and I couldn't get any information from getLastError() or getLastResponse(). I ran the code step by step and found that if NEWSLETTER_ENDPOINT is set to an empty string as the documentation states, the Mailchimp endpoint isn't built correctly. The solution is to set that value to null instead.